+++ NEWS.md
@@ -1,4 +1,7 @@
 # mergen Version 0.1.1 _2023-12-29_
-  * URLs added 
+  * URLs in DESCRIPTION added 
+  * new feature: now users can specify openai-like APIs by providing url and setting
+    agent name as "generic" in setupAgent() function
+  * new feature: clean_code_blocks() now automatically runs when extractCode() is called
   * fix: typos in setupAgent fixed
   * fix: sendPrompt fixed which was broken due to openai API changes